backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
groq>=0.12.0
python-multipart==0.0.6
requests==2.31.0
aiofiles==23.2.1
